+++ R/print.R
@@ -74,7 +74,7 @@ print.apd_hat_values <- function(x, ...) {
 #' tr_x <- matrix(
 #'   sample(0:1, size = 20 * 50, prob = rep(.5, 2), replace = TRUE),
 #'   ncol = 20
-#'  )
+#' )
 #' model <- apd_similarity(tr_x)
 #' print(model)
 #' @export