+++ NEWS.md
@@ -1,9 +1,21 @@
-applicable 0.0.1.1
-==================
+# applicable 0.1.1
+
+
+- Changes in [apd_isolation()] because [isotree::isolation.forest()] changed their default for the number of dimensions to 1. 
+
+- Updated snapshot unit tests that changed from rlang's changes related to the call/traceback. 
+
+# applicable 0.1.0
+
+
+Added isolation forest methods via the isotree package in the function `apd_isolation()`.
+
+# applicable 0.0.1.1
+
 
 Minor patch release: fixed failing units tests due to recent package updates.
 
-applicable 0.0.1
-==================
+# applicable 0.0.1
+
 
 * First CRAN version.